packages/reference/services/demo-python-service/requirements.txt
Flask==3.0.3
Flask-SQLAlchemy==3.1.1
requests==2.31.0
gunicorn==21.2.0
psycopg2-binary==2.9.6
python-dotenv==1.0.0
